+++ packages/components/ui/src/Link.svelte
@@ -1,7 +1,7 @@
 <script>
 	import ExternalLink from './icons/feather/ExternalLink.svelte';
 	import Icon from './icons/Icon.svelte';
-	import {defaultRel} from './utils/shared';
+	import {defaultRel} from './utils/shared.js';
 
 	const defaultIconSize = 14;
 	const {defaultStrokeWidth} = Icon;